need to create a cleanup function
